requirements.txt
streamlit
requests
assemblyai
google-generativeai
streamlit-audio-recorder
